rename from_proto.h -> proto_tree_iterator.h
