+++ searchlib/src/vespa/searchlib/common/serialized_query_tree.cpp
@@ -3,7 +3,7 @@
 #include "serialized_query_tree.h"
 #include <vespa/searchlib/engine/search_protocol_proto.h>
 #include <vespa/searchlib/parsequery/stackdumpiterator.h>
-#include <vespa/searchlib/query/from_proto.h>
+#include <vespa/searchlib/query/proto_tree_iterator.h>
 
 namespace search {
