+++ searchlib/src/vespa/searchlib/query/CMakeLists.txt
@@ -1,9 +1,9 @@
 # Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchlib_query OBJECT
     SOURCES
-    from_proto.cpp
     numeric_range_spec.cpp
     proto_tree_converter.cpp
+    proto_tree_iterator.cpp
     query_normalization.cpp
     query_stack_iterator.cpp
     query_term_decoder.cpp
